requirements.txt
# Core dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pymilvus>=2.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
pdfplumber>=0.10.0
easyocr>=1.7.0
nltk>=3.8.0
requests>=2.31.0

# LangChain dependencies
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
langchain-ollama>=0.0.1
langchain-milvus>=0.0.1
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
pypdf>=3.17.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
